+++ scripts/apps/formula-group-config.mjs
@@ -90,10 +90,26 @@ export class FormulaGroupConfig extends BaseEntitySheet {
 
         formData = expandObject(formData);
 
+        // If there are no formula groups in the form or no formulae to be in groups, quit early
+        if (!formData.formulaGroupLabels?.length || !formData.formulaGroupContains) return;
+
+        // Ensure that things that should be arrays are actually arrays
+        formData.formulaGroupLabels = formData.formulaGroupLabels instanceof Array ? formData.formulaGroupLabels : [formData.formulaGroupLabels];
+        formData.formulaGroupContains = Object.values(formData.formulaGroupContains).map(x => x instanceof Array ? x : [x]);
+
+        // Create formula groups flag from properly formatted form data.
         const formulaGroups = formData.formulaGroupLabels.map((groupLabel, groupIdx) => ({
             label: groupLabel,
             formulaSet: formData.formulaGroupContains[groupIdx].map((x, i) => x ? i : null).filter(x => x != null),
         }));
+
+        // if this change would cause all formula groups to be empty, cancel
+        const allGroupsEmpty = formulaGroups.every(fg => !fg.formulaSet.length);
+        if (allGroupsEmpty) {
+            ui.notifications.warn(game.i18n.localize(`${MODULE_NAME}.FORMULA-GROUP.GroupsMustNotBeEmpty`));
+            this.render(true);
+            return;
+        }
 
         await this.entity.setFlag(MODULE_NAME, "formulaGroups", formulaGroups);
         this.position.width = "auto";